Moved mail model to internal directory
